Restructured Response Mapper
